Cadastrar Produtos completo(web)

+++ ApiStorage/src/main/java/com/FletserTech/storageflet/service/ProductService.java
@@ -7,10 +7,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.FletserTech.storageflet.repository.ProductRepository;
+
+import jakarta.persistence.EntityNotFoundException;
+
 import com.FletserTech.storageflet.models.ProductModel;
 
 @Service
 public class ProductService {
+
     @Autowired
     private ProductRepository productRepository;
 
@@ -23,11 +27,16 @@ public class ProductService {
     }
 
     public ProductModel save(ProductModel product) {
+        if (product.getName() == null || product.getSupplier() == null) {
+            throw new IllegalArgumentException("Product name and supplier must not be null");
+        }
         return productRepository.save(product);
     }
 
     public void deleteById(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new EntityNotFoundException("Product with ID " + id + " does not exist");
+        }
         productRepository.deleteById(id);
     }
-
 }
